modify index.ejs and dashboardcontroller to add to-do status in homepage, add validations in created job and updated job, add validation in updated perfil, add function up job seeing the trade cards to-do and progress, add alerts validation in created job, perfil, update job... add function down job seeing the trade cards to-do and progress, created rule in job to no altered daily-hours in the job progress, altered rule profile and notify user about hours-per-day, modify method calcule DiffDays taking into account the days worked and not worked in the week

+++ src/db/init.js
@@ -8,7 +8,7 @@ const initDb = {
     id INTEGER PRIMARY KEY AUTOINCREMENT,
     name TEXT,
     avatar TEXT,
-    monthly_budget INT,
+    monthly_budget REAL,
     days_per_week INT,
     hours_per_day INT,
     vacation_per_year INT,
